Set user name length and $TERM length to values previously used in Debian
Changes (in os.h) MAXTERMLEN from 32 to 40 and MAXLOGINLEN from 256 to
50, i.e. to those values previously used in  49long-usernames.patch
and 50increase-max-TERM-length.patch to ensure backwards compatibility
in Debian. (See #644788 for the discussion with older releases.)

Also changes MSG_VERSION back to 3 in screen.h.

Also drops patch 60-644788-screen-4.1.0-4.0.3-interoperability.patch
as it was only necessary for the upgrade from Debian 6 Squeeze to
Debian 7 Wheezy and from Ubuntu 12.04 LTS Precise to Ubuntu 14.04 LTS
Trusty.